+++ src/xurrent/skill_pools.py
@@ -3,39 +3,45 @@ from typing import Optional, List, TypeVar
 from .core import XurrentApiHelper, JsonSerializableDict
 from enum import Enum
 
-T = TypeVar('T', bound='Calendar')
+T = TypeVar('T', bound='SkillPool')
 
 
-class CalendarPredefinedFilter(str, Enum):
-    enabled = "enabled"
+class SkillPoolPredefinedFilter(str, Enum):
     disabled = "disabled"
+    enabled = "enabled"
 
     def __str__(self):
         return self.value
 
 
-class Calendar(JsonSerializableDict):
-    # https://developer.xurrent.com/v1/calendars/
-    __resourceUrl__ = 'calendars'
+class SkillPool(JsonSerializableDict):
+    # https://developer.xurrent.com/v1/skill_pools/
+    __resourceUrl__ = 'skill_pools'
 
     def __init__(self,
                  connection_object: XurrentApiHelper,
                  id: int,
                  name: Optional[str] = None,
                  disabled: Optional[bool] = None,
+                 manager=None,
                  **kwargs):
         self.id = id
         self._connection_object = connection_object
         self.name = name
         self.disabled = disabled
+
+        from .people import Person
+        self.manager = (manager if isinstance(manager, Person)
+                        else Person.from_data(connection_object, manager) if manager else None)
+
         for key, value in kwargs.items():
             setattr(self, key, value)
 
     def __str__(self) -> str:
-        return f"Calendar(id={self.id}, name={self.name}, disabled={self.disabled})"
+        return f"SkillPool(id={self.id}, name={self.name}, disabled={self.disabled})"
 
     def ref_str(self) -> str:
-        return f"Calendar(id={self.id}, name={self.name})"
+        return f"SkillPool(id={self.id}, name={self.name})"
 
     @classmethod
     def from_data(cls, connection_object: XurrentApiHelper, data) -> T:
@@ -51,9 +57,9 @@ class Calendar(JsonSerializableDict):
         return cls.from_data(connection_object, connection_object.api_call(uri, 'GET'))
 
     @classmethod
-    def get_calendars(cls, connection_object: XurrentApiHelper,
-                      predefinedFilter: CalendarPredefinedFilter = None,
-                      queryfilter: dict = None) -> List[T]:
+    def get_skill_pools(cls, connection_object: XurrentApiHelper,
+                        predefinedFilter: SkillPoolPredefinedFilter = None,
+                        queryfilter: dict = None) -> List[T]:
         uri = f'{connection_object.base_url}/{cls.__resourceUrl__}'
         if predefinedFilter:
             uri = f'{uri}/{predefinedFilter}'
@@ -65,7 +71,7 @@ class Calendar(JsonSerializableDict):
     def update(self, data: dict) -> T:
         uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}'
         response = self._connection_object.api_call(uri, 'PATCH', data)
-        return Calendar.from_data(self._connection_object, response)
+        return SkillPool.from_data(self._connection_object, response)
 
     @classmethod
     def create(cls, connection_object: XurrentApiHelper, data: dict) -> T:
@@ -79,30 +85,18 @@ class Calendar(JsonSerializableDict):
     def disable(self) -> T:
         return self.update({'disabled': True})
 
-    def get_duration(self, start: str, end: str, time_zone: str = None) -> dict:
-        """
-        Calculate the duration between two timestamps according to the calendar.
-
-        :param start: Start datetime string (ISO 8601)
-        :param end: End datetime string (ISO 8601)
-        :param time_zone: Optional time zone name
-        :return: Duration data from the API
-        """
-        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/duration'
-        params = f'start={start}&end={end}'
-        if time_zone:
-            params += f'&time_zone={time_zone}'
-        uri += f'?{params}'
-        return self._connection_object.api_call(uri, 'GET')
-
-    def get_hours(self) -> List[dict]:
-        """Retrieve the working hours of the calendar."""
-        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/hours'
-        return self._connection_object.api_call(uri, 'GET')
-
-    def get_holidays(self) -> List:
-        """Retrieve the holidays associated with this calendar."""
-        from .holidays import Holiday
-        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/holidays'
+    def get_members(self, queryfilter: dict = None) -> List:
+        from .people import Person
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/members'
+        if queryfilter:
+            uri += '?' + self._connection_object.create_filter_string(queryfilter)
         response = self._connection_object.api_call(uri, 'GET')
-        return [Holiday.from_data(self._connection_object, h) for h in response]
+        return [Person.from_data(self._connection_object, item) for item in response]
+
+    def get_effort_classes(self, queryfilter: dict = None) -> List:
+        from .effort_classes import EffortClass
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/effort_classes'
+        if queryfilter:
+            uri += '?' + self._connection_object.create_filter_string(queryfilter)
+        response = self._connection_object.api_call(uri, 'GET')
+        return [EffortClass.from_data(self._connection_object, item) for item in response]